support multiple options with arguments
For each option that takes an argument, create an additional array
that keeps track of each use of that option. The key of this array
is the character of the option prefixed with an underscore. This
allows the use of multiple invocations of the same option with
different option arguments, i.e. "cmd -q val1 -q val2".